Use Packer Template feature to auto-populate ISOs
Packer template variables allow for dynamic integration of isodate
variables.
This commit leverages the isotime template engine to auto generate
the year and month values for the ISO url as well as the SHA checksum
fields.

This will allow for the ISO image to continuously remain up-to-date
without the need for continued monthly commits.

Additionally, it leverages this same template variable to ensure that
the resulting box filename matches the date it was created, for archival
purposes.